Fix .dockstore.yml
This entry is currently not displaying correctly on Dockstore, likely due to the lack of ".wdl" in this .dockstore.yml:
* [2.1.1](https://dockstore.org/workflows/github.com/gatk-workflows/gatk4-data-processing/processing-for-variant-discovery-gatk4:2.1.1?tab=info)
* [Master](https://dockstore.org/workflows/github.com/gatk-workflows/gatk4-data-processing/processing-for-variant-discovery-gatk4:master?tab=versions)

Older entries such as [2.1.0](https://dockstore.org/workflows/github.com/gatk-workflows/gatk4-data-processing/processing-for-variant-discovery-gatk4:2.1.0?tab=files) are still working.

This PR should fix the issue on master and any subsequent releases. The addition of the `authors` field and the spacing adjustments are recommended but not strictly required.